refactoring: move middleware for users, boards, tasks to files into according folders (users, boards, tasks) from separate files in utils
