+++ src/resources/tasks/task.router.js
@@ -1,9 +1,10 @@
 const router = require('express').Router();
 
-const getTaskId = require('../../utils/get.task.id');
+const taskMiddleware = require('./task.middleware');
+
+// res.locals should contain board already
 
 router.route('/').get(async (req, res) => {
-  // const { board } = res.locals;
   const tasks = [];
 
   res.status(200).json(tasks);
@@ -13,7 +14,7 @@ router.route('/').post(async (req, res) => {
   res.sendStatus(501);
 });
 
-router.use('/:taskId', getTaskId);
+router.use('/:taskId', taskMiddleware.getTaskId);
 
 router.route('/:taskId').get(async (req, res) => {
   res.sendStatus(501);
